chore: move project template classes for module app
