+++ app/src/main/res/layout/layout_project_template.xml
