+++ app/src/main/res/layout/layout_template_view.xml
